requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pandas>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
motor>=3.3.0

aiosmtplib>=3.0.0
